+++ R/pos.r
@@ -19,9 +19,10 @@
 #' Basically, the function will return a list of character vectors with (morpheme)/(tag) elements.
 #'
 #' @param sentence A character vector of any length. For analyzing multiple sentences, put them in one character vector.
+#' @param join A bool to decide the output format. The default value is TRUE. If FALSE, the function will return morphemes only, and tags put in the attribute. if \code{format="data.frame"}, then this will be ignored.
+#' @param format A data type for the result. The default value is "list". You can set this to "data.frame" to get a result as data frame format.
 #' @param sys_dic A location of system MeCab dictionary. The default value is "".
 #' @param user_dic A location of user-specific MeCab dictionary. The default value is "".
-#' @param join A bool to decide the output format. The default value is TRUE. If FALSE, the function will return morphemes only, and tags put in the attribute.
 #' @return A string vector of POS tagged morpheme will be returned in conjoined character
 #'  vecter form. Element name of the list are original phrases
 #'
@@ -30,13 +31,14 @@
 #' sentence <- c(#some UTF-8 texts)
 #' pos(sentence)
 #' pos(sentence, join = FALSE)
+#' pos(sentence, format = "data.frame")
 #' pos(sentence, user_dic = "~/user_dic.dic")
-#' # in case of using mecab-ipadic-NEologd
+#' # System dictionary example: in case of using mecab-ipadic-NEologd
 #' pos(sentence, sys_dic = "/usr/local/lib/mecab/dic/mecab-ipadic-neologd/")
 #' }
 #'
 #' @export
-pos <- function(sentence, join = TRUE, sys_dic = "", user_dic = "") {
+pos <- function(sentence, join = TRUE, format = c("list", "data.frame"), sys_dic = "", user_dic = "") {
   if (typeof(sentence) != "character") {
     if (typeof(sentence) == "factor") {
       stop("The type of input sentence is a factor. Please typesetting it with as.character().")
@@ -47,20 +49,25 @@ pos <- function(sentence, join = TRUE, sys_dic = "", user_dic = "") {
 
   if (!is.null(getOption("mecabSysDic")) && !sys_dic == "") sys_dic = getOption("mecabSysDic")
 
-  if (join == TRUE) {
-    if (length(sentence) > 1) {
-      result <- posLoopJoinRcpp(sentence, sys_dic, user_dic)
+  format = match.arg(format)
+
+  if (format == "data.frame") {
+    result <- posDFRcpp(sentence, sys_dic, user_dic)
+  } else{
+    if (join == TRUE) {
+      if (length(sentence) > 1) {
+        result <- posLoopJoinRcpp(sentence, sys_dic, user_dic)
+      } else {
+        result <- posJoinRcpp(sentence, sys_dic, user_dic)
+      }
     } else {
-      result <- posJoinRcpp(sentence, sys_dic, user_dic)
-    }
-  } else {
-    if (length(sentence) > 1) {
-      result <- posLoopRcpp(sentence, sys_dic, user_dic)
-    } else {
-      result <- posRcpp(sentence, sys_dic, user_dic)
+      if (length(sentence) > 1) {
+        result <- posLoopRcpp(sentence, sys_dic, user_dic)
+      } else {
+        result <- posRcpp(sentence, sys_dic, user_dic)
+      }
     }
   }
-
 
   return(result)
 }