+++ R/posParallel.R
@@ -24,9 +24,10 @@
 #' Basically, the function will return a list of character vectors with (morpheme)/(tag) elements.
 #'
 #' @param sentence A character vector of any length. For analyzing multiple sentences, put them in one character vector.
+#' @param join A bool to decide the output format. The default value is TRUE. If FALSE, the function will return morphemes only, and tags put in the attribute. if \code{format="data.frame"}, then this will be ignored.
+#' @param format A data type for the result. The default value is "list". You can set this to "data.frame" to get a result as data frame format.
 #' @param sys_dic A location of system MeCab dictionary. The default value is "".
 #' @param user_dic A location of user-specific MeCab dictionary. The default value is "".
-#' @param join A bool to decide the output format. The default value is TRUE. If FALSE, the function will return morphemes only, and tags put in the attribute.
 #' @return A string vector of POS tagged morpheme will be returned in conjoined character
 #'  vecter form. Element name of the list are original phrases
 #'
@@ -35,13 +36,14 @@
 #' sentence <- c(#some UTF-8 texts)
 #' posParallel(sentence)
 #' posParallel(sentence, join = FALSE)
+#' posParallel(sentence, format = "data.frame")
 #' posParallel(sentence, user_dic = "~/user_dic.dic")
-#' # in case of using mecab-ipadic-NEologd
+#' # System dictionary example: in case of using mecab-ipadic-NEologd
 #' pos(sentence, sys_dic = "/usr/local/lib/mecab/dic/mecab-ipadic-neologd/")
 #' }
 #'
 #' @export
-posParallel <- function(sentence, join = TRUE, sys_dic = "", user_dic = "") {
+posParallel <- function(sentence, join = TRUE, format = c("list", "data.frame"), sys_dic = "", user_dic = "") {
   if (typeof(sentence) != "character") {
     if (typeof(sentence) == "factor") {
       stop("The type of input sentence is a factor. Please typesetting it with as.character().")
@@ -50,12 +52,18 @@ posParallel <- function(sentence, join = TRUE, sys_dic = "", user_dic = "") {
     }
   }
 
+  format = match.arg(format)
+
   if (!is.null(getOption("mecabSysDic")) && !sys_dic == "") sys_dic = getOption("mecabSysDic")
 
-  if (join == TRUE) {
-    result <- posParallelJoinRcpp(sentence, sys_dic, user_dic)
+  if (format == "data.frame") {
+    result <- posParallelDFRcpp(sentence, sys_dic, user_dic)
   } else {
-    result <- posParallelRcpp(sentence, sys_dic, user_dic)
+    if (join == TRUE) {
+      result <- posParallelJoinRcpp(sentence, sys_dic, user_dic)
+    } else {
+      result <- posParallelRcpp(sentence, sys_dic, user_dic)
+    }
   }
 
   return(result)