+++ src/RcppExports.cpp
@@ -31,6 +31,19 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// posDFRcpp
+DataFrame posDFRcpp(StringVector text, std::string sys_dic, std::string user_dic);
+RcppExport SEXP _RcppMeCab_posDFRcpp(SEXP textSEXP, SEXP sys_dicSEXP, SEXP user_dicSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< StringVector >::type text(textSEXP);
+    Rcpp::traits::input_parameter< std::string >::type sys_dic(sys_dicSEXP);
+    Rcpp::traits::input_parameter< std::string >::type user_dic(user_dicSEXP);
+    rcpp_result_gen = Rcpp::wrap(posDFRcpp(text, sys_dic, user_dic));
+    return rcpp_result_gen;
+END_RCPP
+}
 // posParallelJoinRcpp
 List posParallelJoinRcpp(std::vector<std::string> text, std::string sys_dic, std::string user_dic);
 RcppExport SEXP _RcppMeCab_posParallelJoinRcpp(SEXP textSEXP, SEXP sys_dicSEXP, SEXP user_dicSEXP) {
@@ -41,6 +54,19 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< std::string >::type sys_dic(sys_dicSEXP);
     Rcpp::traits::input_parameter< std::string >::type user_dic(user_dicSEXP);
     rcpp_result_gen = Rcpp::wrap(posParallelJoinRcpp(text, sys_dic, user_dic));
+    return rcpp_result_gen;
+END_RCPP
+}
+// posParallelDFRcpp
+DataFrame posParallelDFRcpp(StringVector text, std::string sys_dic, std::string user_dic);
+RcppExport SEXP _RcppMeCab_posParallelDFRcpp(SEXP textSEXP, SEXP sys_dicSEXP, SEXP user_dicSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< StringVector >::type text(textSEXP);
+    Rcpp::traits::input_parameter< std::string >::type sys_dic(sys_dicSEXP);
+    Rcpp::traits::input_parameter< std::string >::type user_dic(user_dicSEXP);
+    rcpp_result_gen = Rcpp::wrap(posParallelDFRcpp(text, sys_dic, user_dic));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -87,7 +113,9 @@ END_RCPP
 static const R_CallMethodDef CallEntries[] = {
     {"_RcppMeCab_posLoopRcpp", (DL_FUNC) &_RcppMeCab_posLoopRcpp, 3},
     {"_RcppMeCab_posLoopJoinRcpp", (DL_FUNC) &_RcppMeCab_posLoopJoinRcpp, 3},
+    {"_RcppMeCab_posDFRcpp", (DL_FUNC) &_RcppMeCab_posDFRcpp, 3},
     {"_RcppMeCab_posParallelJoinRcpp", (DL_FUNC) &_RcppMeCab_posParallelJoinRcpp, 3},
+    {"_RcppMeCab_posParallelDFRcpp", (DL_FUNC) &_RcppMeCab_posParallelDFRcpp, 3},
     {"_RcppMeCab_posParallelRcpp", (DL_FUNC) &_RcppMeCab_posParallelRcpp, 3},
     {"_RcppMeCab_posRcpp", (DL_FUNC) &_RcppMeCab_posRcpp, 3},
     {"_RcppMeCab_posJoinRcpp", (DL_FUNC) &_RcppMeCab_posJoinRcpp, 3},